+++ client/src/components/Login/LoginValidationLogic/LoginValidationLogic.tsx
@@ -1,7 +1,6 @@
-import { ValidityResults } from '../../../pages/Login/Login';
+import { ValidityResults } from '../LoginSection/LoginSection';
 import { LoginValue } from '../../../api/Login';
 
-// 이메일 정규 표현식 => 일치 시 true 반환
 export const validateEmail = (email: string) => {
   const regExp =
     /^[0-9a-zA-Z]([-_\.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_\.]?[0-9a-zA-Z])*\.[a-zA-Z]{2,3}$/i;
@@ -9,7 +8,6 @@ export const validateEmail = (email: string) => {
   return regExp.test(email);
 };
 
-// 비밀번호 정규 표현식 => 일치 시 true 반환
 //? 8 ~ 20자 영문, 숫자, 특수문자 최소 한가지씩 조합
 export const validatePassword = (password: string) => {
   const regExp =
